+++ .env
@@ -1,1 +1,4 @@
-GEMINI_API_KEY ="Ayour api key"
+GEMINI_API_KEY=AIzaSyCXvGEt7fgYfeeTXi6idvcGLYyYRd7obuw
+GEMINI_API_KEY1=AIzaSyCnNLp6j8j4OEpPOUo53n34QG2P0QeECS0
+GEMINI_API_KEY2=AIzaSyBZtVuOsoc4ry4aSmcokG0RPFrrSfGqVzQ
+GEMINI_API_KEY4=AIzaSyD3Gx75zv_HrzXVO3_ei-N_geZCKcaCcyE
