--- tools/__init__.py
@@ -1,1 +1,0 @@
-from tools.tools import Api
